Lesson 7 Spring Data Homework Работа на ошибками Исправление ошибки с фильтрами

+++ lesson-04/src/main/java/ru/geekbrain/lesson04/controller/ProductController.java
@@ -38,7 +38,7 @@ public class ProductController {
    * Уровень контроллера. Обрабатываем метод GET URL  .../product/*
    */
   @GetMapping
-  public String listPage(@RequestParam Optional<Float> minCost, @RequestParam Optional<Float> maxCost, Model model) {
+  public String listPage(@RequestParam Optional<BigDecimal> minCost, @RequestParam Optional<BigDecimal> maxCost, Model model) {
     // Модель
     model.addAttribute("products", productRepository.findProductByCostBetween(minCost.isPresent() ? minCost.get() : null,
             maxCost.isPresent() ? maxCost.get() : null));
